core: narrow arc invalidation to a grid of cells
A Path's bounding rect is its whole element, and an arc gauge fills the element
it lives in, so moving the sweep by a fraction of a degree invalidated the
entire panel and every label on it redrew. Measured on hardware at 100% dirty:
prepare 30ms, render 40ms, flush 10ms. Render and flush both scale with the
dirty area, so this was 50 of 80ms.

Divide the element into a 5x5 grid and round invalidation out to whole cells.
Cells are ~93px across, and rounding outward can only ever repaint too much,
never too little. That matters more than the precision it gives up: an earlier
version computed the exact swept difference and spent four attempts failing to
get its last pixel right, because where an arc runs parallel to the region's
edge - by the horizontal axes - one lost pixel column costs a wide band of arc.
Cells also coalesce where exact bands multiplied, which counts because a
DirtyRegion holds three rects before merging them into their union.

Track the owed cells as a bit mask, accumulated as the arc changes and cleared
only when it is drawn. A region computed but never painted keeps its debt
instead of being forgotten, which is how the exact version lost track of what
was actually on screen.

Adds tests at the size a dial is really drawn - 466px with a 12px stroke, where
most rows pass through the ring's hole and split into two runs, which a 22px
test arc barely exercises. They cover the sectors either side of both
horizontal axes, that no row a sector reaches is left unpainted, and that
nothing is painted outside the sector; the last of those checks over-painting,
which the others structurally cannot see.

+++ internal/renderers/software/scene.rs
@@ -531,10 +531,16 @@ pub struct ArcCommand {
     pub outer_radius: PhysicalLength,
     pub inner_radius: PhysicalLength,
     pub color: PremultipliedRgbaColor,
-    /// Unit vectors along the two boundary rays, in 1.15 fixed point. A point is inside
-    /// the wedge when it is on the inner side of both (or, for a reflex sweep, either).
-    pub start_dir: (i32, i32),
-    pub end_dir: (i32, i32),
+    /// Unit vectors along the two boundary rays. A point is inside the wedge when it is on
+    /// the inner side of both (or, for a reflex sweep, either).
+    ///
+    /// Kept in f32 rather than fixed point on purpose: the row's clip bound divides by the
+    /// ray's y component, so for a near horizontal ray any error in it is amplified by
+    /// 1/y. At a tenth of a degree from horizontal, 1.15 fixed point was enough to move the
+    /// bound by over a hundred pixels, which showed up as the arc failing to draw around 3
+    /// and 9 o'clock.
+    pub start_dir: (f32, f32),
+    pub end_dir: (f32, f32),
     /// Sweep greater than 180 degrees, where the wedge is the union of the two half
     /// planes rather than their intersection.
     pub reflex: bool,
